+++ ChangeLog.md
@@ -1,6 +1,10 @@
 ### 0.18.0
 - Added: MinorArc::relate and Loop::relate
+- Added: Rectangle::intersects
+- Added: Cap:intersects
+- Added: Cap:expand
 - Added: direct/inverse rotation matrix and origin of local frame
+- Fixed: Cap::contains_cap
 
 ### 0.17.0
 - Added: geo-types and geo-traits (behind a feature flag)